Refactor CLI commands into w/ project and w/o project
