--- datumaro/cli/commands/__init__.py
@@ -1,31 +1,0 @@
-# Copyright (C) 2019-2022 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-# pylint: disable=redefined-builtin
-
-from . import (
-    add,
-    checkout,
-    commit,
-    convert,
-    create,
-    detect_format,
-    diff,
-    explain,
-    export,
-    filter,
-    generate,
-    import_,
-    info,
-    log,
-    merge,
-    patch,
-    remove,
-    search,
-    stats,
-    status,
-    transform,
-    validate,
-)
-from .no_project import download
