--- datumaro/cli/commands/export.py
@@ -1,11 +1,0 @@
-# Copyright (C) 2019-2021 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-from ..contexts.project import build_export_parser as build_parser
-from ..contexts.project import get_export_sensitive_args as get_sensitive_args
-
-__all__ = [
-    "build_parser",
-    "get_sensitive_args",
-]
