--- datumaro/cli/commands/import_.py
@@ -1,11 +1,0 @@
-# Copyright (C) 2020-2021 Intel Corporation
-#
-# SPDX-License-Identifier: MIT
-
-from ..contexts.source import build_import_parser as build_parser
-from ..contexts.source import get_import_sensitive_args as get_sensitive_args
-
-__all__ = [
-    "build_parser",
-    "get_sensitive_args",
-]
