+++ datumaro/cli/contexts/project/__init__.py
@@ -18,291 +18,11 @@ from datumaro.util import dump_json_file, str_to_bool
 from datumaro.util.os_util import make_file_name
 from datumaro.util.scope import scope_add, scoped
 
-from ...util import MultilineFormatter, add_subparser
+from ...project_commands import get_project_commands
+from ...project_commands.dataset_operations import export, stats
+from ...util import MultilineFormatter, add_subparser, make_subcommands_help
 from ...util.errors import CliException
 from ...util.project import generate_next_file_name, load_project, parse_full_revpath
-
-
-class FilterModes(Enum):
-    # primary
-    items = 1
-    annotations = 2
-    items_annotations = 3
-
-    # shortcuts
-    i = 1
-    a = 2
-    i_a = 3
-    a_i = 3
-    annotations_items = 3
-
-    @staticmethod
-    def parse(s):
-        s = s.lower()
-        s = s.replace("+", "_")
-        return FilterModes[s]
-
-    @classmethod
-    def make_filter_args(cls, mode):
-        if mode == cls.items:
-            return {}
-        elif mode == cls.annotations:
-            return {"filter_annotations": True}
-        elif mode == cls.items_annotations:
-            return {
-                "filter_annotations": True,
-                "remove_empty": True,
-            }
-        else:
-            raise NotImplementedError()
-
-    @classmethod
-    def list_options(cls):
-        return [m.name.replace("_", "+") for m in cls]
-
-
-def build_export_parser(parser_ctor=argparse.ArgumentParser):
-    builtins = sorted(Environment().exporters)
-
-    parser = parser_ctor(
-        help="Export project",
-        description="""
-        Exports a project in some format.|n
-        |n
-        Each dataset format has its own export
-        options, which are passed after the '--' separator (see examples),
-        pass '-- -h' for more info. If not stated otherwise, by default
-        only annotations are exported, to include images pass
-        '--save-images' parameter.|n
-        |n
-        A filter can be passed, check the 'filter' command description for
-        more info.|n
-        |n
-        Formats:|n
-        Datasets come in a wide variety of formats. Each dataset
-        format defines its own data structure and rules on how to
-        interpret the data. Check the user manual for the list of
-        supported formats, examples and documentation.
-        |n
-        The list of supported formats can be extended by plugins.
-        Check the "plugins" section of the developer guide for information
-        about plugin implementation.|n
-        |n
-        List of builtin dataset formats: {}|n
-        |n
-        The command can only be applied to a project build target, a stage
-        or the combined 'project' target, in which case all the targets will
-        be affected.
-        |n
-        Examples:|n
-        - Export project as a VOC-like dataset, include images:|n
-        |s|s%(prog)s -f voc -- --save-images|n
-        |n
-        - Export project as a COCO-like dataset in other directory:|n
-        |s|s%(prog)s -f coco -o path/I/like/
-        """.format(
-            ", ".join(builtins)
-        ),
-        formatter_class=MultilineFormatter,
-    )
-
-    parser.add_argument(
-        "_positionals", nargs=argparse.REMAINDER, help=argparse.SUPPRESS
-    )  # workaround for -- eaten by positionals
-    parser.add_argument(
-        "target",
-        nargs="?",
-        default="project",
-        help="A project target to be exported (default: %(default)s)",
-    )
-    parser.add_argument("-e", "--filter", help="XML XPath filter expression for dataset items")
-    parser.add_argument(
-        "--filter-mode",
-        default=FilterModes.i.name,
-        type=FilterModes.parse,
-        help="Filter mode (options: %s; default: %s)"
-        % (", ".join(FilterModes.list_options()), "%(default)s"),
-    )
-    parser.add_argument(
-        "-o",
-        "--output-dir",
-        dest="dst_dir",
-        help="Directory to save output (default: a subdir in the current one)",
-    )
-    parser.add_argument(
-        "--overwrite", action="store_true", help="Overwrite existing files in the save directory"
-    )
-    parser.add_argument(
-        "-p",
-        "--project",
-        dest="project_dir",
-        help="Directory of the project to operate on (default: current dir)",
-    )
-    parser.add_argument("-f", "--format", required=True, help="Output format")
-    parser.add_argument(
-        "extra_args",
-        nargs=argparse.REMAINDER,
-        help="Additional arguments for exporter (pass '-- -h' for help). "
-        "Must be specified after the main command arguments and after "
-        "the '--' separator",
-    )
-    parser.set_defaults(command=export_command)
-
-    return parser
-
-
-def get_export_sensitive_args():
-    return {
-        export_command: ["dst_dir", "project_dir", "name", "extra_args", "target", "filter"],
-    }
-
-
-@scoped
-def export_command(args):
-    has_sep = "--" in args._positionals
-    if has_sep:
-        pos = args._positionals.index("--")
-        if 1 < pos:
-            raise argparse.ArgumentError(None, message="Expected no more than 1 target argument")
-    else:
-        pos = 1
-    args.target = (args._positionals[:pos] or [ProjectBuildTargets.MAIN_TARGET])[0]
-    args.extra_args = args._positionals[pos + has_sep :]
-
-    show_plugin_help = "-h" in args.extra_args or "--help" in args.extra_args
-
-    project = None
-    try:
-        project = scope_add(load_project(args.project_dir))
-    except ProjectNotFoundError:
-        if not show_plugin_help:
-            raise
-
-    if project is not None:
-        env = project.env
-    else:
-        env = Environment()
-
-    try:
-        exporter = env.exporters[args.format]
-    except KeyError:
-        raise CliException("Exporter for format '%s' is not found" % args.format)
-
-    extra_args = exporter.parse_cmdline(args.extra_args)
-
-    dst_dir = args.dst_dir
-    if dst_dir:
-        if not args.overwrite and osp.isdir(dst_dir) and os.listdir(dst_dir):
-            raise CliException(
-                "Directory '%s' already exists " "(pass --overwrite to overwrite)" % dst_dir
-            )
-    else:
-        dst_dir = generate_next_file_name("export-%s" % make_file_name(args.format))
-    dst_dir = osp.abspath(dst_dir)
-
-    if args.filter:
-        filter_args = FilterModes.make_filter_args(args.filter_mode)
-        filter_expr = args.filter
-
-    log.info("Loading the project...")
-
-    dataset = project.working_tree.make_dataset(args.target)
-    if args.filter:
-        dataset.filter(filter_expr, **filter_args)
-
-    log.info("Exporting...")
-
-    dataset.export(save_dir=dst_dir, format=exporter, **extra_args)
-
-    log.info("Results have been saved to '%s'" % dst_dir)
-
-    return 0
-
-
-def build_stats_parser(parser_ctor=argparse.ArgumentParser):
-    parser = parser_ctor(
-        help="Get project statistics",
-        description="""
-        Outputs various project statistics like image mean and std,
-        annotations count etc.|n
-        |n
-        Target dataset is specified by a revpath. The full syntax is:|n
-        - Dataset paths:|n
-        |s|s- <dataset path>[ :<format> ]|n
-        - Revision paths:|n
-        |s|s- <project path> [ @<rev> ] [ :<target> ]|n
-        |s|s- <rev> [ :<target> ]|n
-        |s|s- <target>|n
-        |n
-        Both forms use the -p/--project as a context for plugins. It can be
-        useful for dataset paths in targets. When not specified, the current
-        project's working tree is used.|n
-        |n
-        Examples:|n
-        - Compute project statistics:|n
-        |s|s%(prog)s
-        """,
-        formatter_class=MultilineFormatter,
-    )
-
-    parser.add_argument(
-        "target", default="project", nargs="?", help="Target dataset revpath (default: project)"
-    )
-    parser.add_argument("-s", "--subset", help="Compute stats only for a specific subset")
-    parser.add_argument(
-        "--image-stats",
-        type=str_to_bool,
-        default=True,
-        help="Compute image mean and std (default: %(default)s)",
-    )
-    parser.add_argument(
-        "--ann-stats",
-        type=str_to_bool,
-        default=True,
-        help="Compute annotation statistics (default: %(default)s)",
-    )
-    parser.add_argument(
-        "-p",
-        "--project",
-        dest="project_dir",
-        help="Directory of the project to operate on (default: current dir)",
-    )
-    parser.set_defaults(command=stats_command)
-
-    return parser
-
-
-def get_stats_sensitive_args():
-    return {
-        stats_command: ["project_dir", "target"],
-    }
-
-
-@scoped
-def stats_command(args):
-    project = None
-    try:
-        project = scope_add(load_project(args.project_dir))
-    except ProjectNotFoundError:
-        if args.project_dir:
-            raise
-
-    dataset, target_project = parse_full_revpath(args.target, project)
-    if target_project:
-        scope_add(target_project)
-
-    if args.subset:
-        dataset = dataset.get_subset(args.subset)
-
-    stats = {}
-    if args.image_stats:
-        stats.update(compute_image_statistics(dataset))
-    if args.ann_stats:
-        stats.update(compute_ann_statistics(dataset))
-
-    dst_file = generate_next_file_name("statistics", ext=".json")
-    log.info("Writing project statistics to '%s'" % dst_file)
-    dump_json_file(dst_file, stats, indent=True)
 
 
 def build_info_parser(parser_ctor=argparse.ArgumentParser):
@@ -469,10 +189,29 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         formatter_class=MultilineFormatter,
     )
 
-    subparsers = parser.add_subparsers()
-    add_subparser(subparsers, "info", build_info_parser)
-    add_subparser(subparsers, "stats", build_stats_parser)
-    add_subparser(subparsers, "export", build_export_parser)
+    subcommands_desc = ""
+    known_commands = get_project_commands()
+    help_line_start = max((len(e[0]) for e in known_commands), default=0)
+    help_line_start = max((2 + help_line_start) // 4 + 1, 6) * 4  # align to tabs
+    if known_commands:
+        if subcommands_desc:
+            subcommands_desc += "\n"
+        subcommands_desc += make_subcommands_help(known_commands, help_line_start)
+    if subcommands_desc:
+        subcommands_desc += (
+            "\nRun '%s COMMAND --help' for more information on a command." % parser.prog
+        )
+
+    # TODO: revisit project related commands for better code structure
+    subparsers = parser.add_subparsers(
+        title=subcommands_desc, description="", help=argparse.SUPPRESS
+    )
+
+    commands = get_project_commands()
+    for command_name, command, _ in commands:
+        if command is not None:
+            add_subparser(subparsers, command_name, command.build_parser)
+
     add_subparser(subparsers, "migrate", build_migrate_parser)
 
     return parser
@@ -481,7 +220,7 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
 def get_sensitive_args():
     return {
         **get_info_sensitive_args(),
-        **get_stats_sensitive_args(),
-        **get_export_sensitive_args(),
+        **stats.get_sensitive_args(),
+        **export.get_sensitive_args(),
         **get_migrate_sensitive_args(),
     }
