+++ datumaro/cli/non_project_commands/convert.py
@@ -11,10 +11,9 @@ from datumaro.components.dataset import Dataset
 from datumaro.components.project import Environment
 from datumaro.util.os_util import make_file_name
 
-from ..contexts.project import FilterModes
 from ..util import MultilineFormatter
 from ..util.errors import CliException
-from ..util.project import generate_next_file_name
+from ..util.project import FilterModes, generate_next_file_name
 
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
