+++ datumaro/cli/non_project_commands/detect_format.py
