+++ datumaro/cli/non_project_commands/filter.py
@@ -7,24 +7,22 @@ import logging as log
 import os
 import os.path as osp
 
-from datumaro.components.environment import Environment
 from datumaro.components.errors import ProjectNotFoundError
+from datumaro.components.filter import DatasetItemEncoder
 from datumaro.components.project import ProjectBuildTargets
 from datumaro.util import str_to_bool
 from datumaro.util.scope import scope_add, scoped
 
 from ..util import MultilineFormatter
 from ..util.errors import CliException
-from ..util.project import load_project, parse_full_revpath
+from ..util.project import FilterModes, load_project, parse_full_revpath
 
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
-    builtins = sorted(Environment().transforms)
-
     parser = parser_ctor(
-        help="Transform project",
+        help="Extract subdataset",
         description="""
-        Applies a batch operation to a dataset and produces a new dataset.|n
+        Extracts a subdataset that contains only items matching filter.|n
         |n
         By default, datasets are updated in-place. The '-o/--output-dir'
         option can be used to specify another output directory. When
@@ -32,7 +30,19 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         updates fail by default to prevent data loss), unless a project
         target is modified.|n
         |n
-        Builtin transforms: {}|n
+        A filter is an XPath expression, which is applied to XML
+        representation of a dataset item. Check '--dry-run' parameter
+        to see XML representations of the dataset items.|n
+        |n
+        To filter annotations use the mode ('-m') parameter.|n
+        Supported modes:|n
+        - 'i', 'items'|n
+        - 'a', 'annotations'|n
+        - 'i+a', 'a+i', 'items+annotations', 'annotations+items'|n
+        When filtering annotations, use the 'items+annotations'
+        mode to point that annotation-less dataset items should be
+        removed. To select an annotation, write an XPath that
+        returns 'annotation' elements (see examples).|n
         |n
         This command has the following invocation syntax:
         - %(prog)s <target dataset revpath>|n
@@ -57,32 +67,43 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         saved if '--apply' is enabled.|n
         |n
         Examples:|n
-        - Convert instance polygons to masks:|n |n
-        |s|s%(prog)s -t polygons_to_masks|n
+        - Filter images with width < height:|n
+        |s|s%(prog)s -e '/item[image/width < image/height]'|n
         |n
-        - Rename dataset items by a regular expression:|n |n
-        |s|s- Replace 'pattern' with 'replacement':|n |n
-        |s|s|s|s%(prog)s -t rename -- -e '|pattern|replacement|'|n
+        - Filter images with large-area bboxes:|n
+        |s|s%(prog)s -e '/item[annotation/type="bbox" and
+            annotation/area>2000]'|n
         |n
-        |s|s- Remove 'frame_' from item ids:|n |n
-        |s|s|s|s%(prog)s -t rename -- -e '|frame_(\\d+)|\\1|'|n
+        - Filter out all irrelevant annotations from items:|n
+        |s|s%(prog)s -m a -e '/item/annotation[label = "person"]'|n
         |n
-        - Split a dataset randomly:|n |n
-        |s|s%(prog)s -t random_split --overwrite path/to/dataset:voc
-        """.format(
-            ", ".join(builtins)
-        ),
+        - Filter out all irrelevant annotations from items:|n
+        |s|s%(prog)s -m a -e '/item/annotation[label="cat" and
+        area > 99.5]'|n
+        |n
+        - Filter occluded annotations and items, if no annotations left:|n
+        |s|s%(prog)s -m i+a -e '/item/annotation[occluded="True"]'|n
+        |n
+        - Filter a VOC-like dataset inplace:|n
+        |s|s%(prog)s -e '/item/annotation[label = "bus"]' --overwrite dataset/:voc
+        """,
         formatter_class=MultilineFormatter,
     )
 
     parser.add_argument(
-        "_positionals", nargs=argparse.REMAINDER, help=argparse.SUPPRESS
-    )  # workaround for -- eaten by positionals
+        "target", nargs="?", default="project", help="Target dataset revpath (default: %(default)s)"
+    )
+    parser.add_argument("-e", "--filter", help="XML XPath filter expression for dataset items")
     parser.add_argument(
-        "target", nargs="?", default="project", help="Target dataset revpath (default: project)"
+        "-m",
+        "--mode",
+        default=FilterModes.i.name,
+        type=FilterModes.parse,
+        help="Filter mode (options: %s; default: %s)"
+        % (", ".join(FilterModes.list_options()), "%(default)s"),
     )
     parser.add_argument(
-        "-t", "--transform", required=True, help="Transform to apply to the dataset"
+        "--dry-run", action="store_true", help="Print XML representations to be filtered and exit"
     )
     parser.add_argument(
         "-o",
@@ -94,15 +115,6 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
             intermediate stages) and dataset targets.
             If not specified, the results will be saved inplace.
             """,
-    )
-    parser.add_argument(
-        "--overwrite", action="store_true", help="Overwrite existing files in the save directory"
-    )
-    parser.add_argument(
-        "-p",
-        "--project",
-        dest="project_dir",
-        help="Directory of the project to operate on (default: current dir)",
     )
     parser.add_argument(
         "--stage",
@@ -125,55 +137,57 @@ def build_parser(parser_ctor=argparse.ArgumentParser):
         "build tree stage will be written (default: %(default)s)",
     )
     parser.add_argument(
-        "extra_args",
-        nargs=argparse.REMAINDER,
-        help="Additional arguments for transformation (pass '-- -h' for help). "
-        "Must be specified after the main command arguments and after "
-        "the '--' separator",
+        "--overwrite", action="store_true", help="Overwrite existing files in the save directory"
     )
-    parser.set_defaults(command=transform_command)
+    parser.add_argument(
+        "-p",
+        "--project",
+        dest="project_dir",
+        help="Directory of the project to operate on (default: current dir)",
+    )
+    parser.set_defaults(command=filter_command)
 
     return parser
 
 
 def get_sensitive_args():
     return {
-        transform_command: ["dst_dir", "project_dir", "extra_args", "target"],
+        filter_command: [
+            "target",
+            "filter",
+            "dst_dir",
+            "project_dir",
+        ],
     }
 
 
 @scoped
-def transform_command(args):
-    has_sep = "--" in args._positionals
-    if has_sep:
-        pos = args._positionals.index("--")
-        if 1 < pos:
-            raise argparse.ArgumentError(None, message="Expected no more than 1 target argument")
-    else:
-        pos = 1
-    args.target = (args._positionals[:pos] or [ProjectBuildTargets.MAIN_TARGET])[0]
-    args.extra_args = args._positionals[pos + has_sep :]
-
-    show_plugin_help = "-h" in args.extra_args or "--help" in args.extra_args
-
+def filter_command(args):
     project = None
     try:
         project = scope_add(load_project(args.project_dir))
     except ProjectNotFoundError:
-        if not show_plugin_help and args.project_dir:
+        if args.project_dir:
             raise
 
-    if project is not None:
-        env = project.env
-    else:
-        env = Environment()
+    filter_args = FilterModes.make_filter_args(args.mode)
+    filter_expr = args.filter
 
-    try:
-        transform = env.transforms[args.transform]
-    except KeyError:
-        raise CliException("Transform '%s' is not found" % args.transform)
+    if args.dry_run:
+        dataset, _project = parse_full_revpath(args.target, project)
+        if _project:
+            scope_add(_project)
 
-    extra_args = transform.parse_cmdline(args.extra_args)
+        dataset = dataset.filter(expr=filter_expr, **filter_args)
+
+        for item in dataset:
+            encoded_item = DatasetItemEncoder.encode(item, dataset.categories())
+            xml_item = DatasetItemEncoder.to_string(encoded_item)
+            print(xml_item)
+        return 0
+
+    if not args.filter:
+        raise CliException("Expected a filter expression ('-e' argument)")
 
     is_target = project is not None and args.target in project.working_tree.build_targets
     if is_target:
@@ -193,17 +207,17 @@ def transform_command(args):
 
         build_tree = project.working_tree.clone()
         for target in targets:
-            build_tree.build_targets.add_transform_stage(target, args.transform, params=extra_args)
+            build_tree.build_targets.add_filter_stage(target, expr=filter_expr, params=filter_args)
 
     if args.apply:
-        log.info("Transforming...")
+        log.info("Filtering...")
 
         if is_target and not args.dst_dir:
             for target in targets:
                 dataset = project.working_tree.make_dataset(build_tree.make_pipeline(target))
 
                 # Source might be missing in the working dir, so we specify
-                # the output directory
+                # the output directory.
                 # We specify save_media here as a heuristic. It can probably
                 # be improved by checking if there are images in the dataset
                 # directory.
@@ -222,7 +236,7 @@ def transform_command(args):
                 )
             dst_dir = osp.abspath(dst_dir)
 
-            dataset.transform(args.transform, **extra_args)
+            dataset.filter(filter_expr, *filter_args)
             dataset.save(dst_dir, save_media=True)
 
             log.info("Results have been saved to '%s'" % dst_dir)
