+++ datumaro/cli/non_project_commands/generate.py
