+++ datumaro/cli/project_commands/dataset_operations/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2023 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from . import diff, export, info, stats, transform, validate
