+++ datumaro/cli/project_commands/dataset_operations/info.py
@@ -8,8 +8,8 @@ from datumaro.components.annotation import AnnotationType
 from datumaro.components.errors import DatasetMergeError, MissingObjectError, ProjectNotFoundError
 from datumaro.util.scope import scope_add, scoped
 
-from ..util import MultilineFormatter
-from ..util.project import load_project, parse_full_revpath
+from ...util import MultilineFormatter
+from ...util.project import load_project, parse_full_revpath
 
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
