+++ datumaro/cli/project_commands/modification/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2023 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from . import add, create, import_, remove
