+++ datumaro/cli/project_commands/versioning/__init__.py
@@ -1,3 +1,5 @@
 # Copyright (C) 2023 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
+
+from . import checkout, commit, log, status
