+++ datumaro/cli/project_commands/versioning/checkout.py
@@ -6,8 +6,8 @@ import argparse
 
 from datumaro.util.scope import scope_add, scoped
 
-from ..util import MultilineFormatter
-from ..util.project import load_project
+from ...util import MultilineFormatter
+from ...util.project import load_project
 
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
