+++ datumaro/cli/project_commands/versioning/status.py
@@ -7,7 +7,7 @@ import argparse
 from datumaro.cli.util import MultilineFormatter
 from datumaro.util.scope import scope_add, scoped
 
-from ..util.project import load_project
+from ...util.project import load_project
 
 
 def build_parser(parser_ctor=argparse.ArgumentParser):
