+++ datumaro/cli/util/project.py
@@ -4,6 +4,7 @@
 
 import os
 import re
+from enum import Enum
 from typing import Optional, Tuple
 
 from datumaro.cli.util.errors import WrongRevpathError
@@ -166,3 +167,41 @@ def split_local_revpath(revpath: str) -> Tuple[Revision, str]:
         target = revpath
 
     return rev, target
+
+
+class FilterModes(Enum):
+    # primary
+    items = 1
+    annotations = 2
+    items_annotations = 3
+
+    # shortcuts
+    i = 1
+    a = 2
+    i_a = 3
+    a_i = 3
+    annotations_items = 3
+
+    @staticmethod
+    def parse(s):
+        s = s.lower()
+        s = s.replace("+", "_")
+        return FilterModes[s]
+
+    @classmethod
+    def make_filter_args(cls, mode):
+        if mode == cls.items:
+            return {}
+        elif mode == cls.annotations:
+            return {"filter_annotations": True}
+        elif mode == cls.items_annotations:
+            return {
+                "filter_annotations": True,
+                "remove_empty": True,
+            }
+        else:
+            raise NotImplementedError()
+
+    @classmethod
+    def list_options(cls):
+        return [m.name.replace("_", "+") for m in cls]
