+++ tests/integration/cli/test_diff.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from datumaro.cli.contexts.project.diff import DiffVisualizer
+from datumaro.cli.util.diff import DiffVisualizer
 from datumaro.components.annotation import (
     AnnotationType,
     Bbox,
@@ -220,6 +220,7 @@ class DiffTest(TestCase):
             result_dir = osp.join(test_dir, "cmp_result")
             run(
                 self,
+                "project",
                 "diff",
                 dataset1_url + ":coco",
                 dataset2_url + ":voc",
@@ -272,6 +273,7 @@ class DiffTest(TestCase):
             result_dir = osp.join(test_dir, "cmp_result")
             run(
                 self,
+                "project",
                 "diff",
                 dataset1_url + ":coco",
                 dataset2_url + ":voc",
